fastboot-protocol: Improve data downloading
When downloading data some fastboot implementation seem to assume full
usb packets are used for everything apart from the last data packet.
Restructure the data download helper to enforce that.

Also extend the example client such that it support android sparse
images as input as well as splitting the input into multiple sparse
images as necessary when downloading to a device