fastboot-protocol: Handle response data with null bytes
There are cases, e.g. Spacemit K1 Bananapi F3 BootROM firmware, where
the fastboot implementation can respond with a message body data that
needs to be parsed as a NULL-terminated string, despite the bytes length
of the response.

Following the protocol message description [0], this commit makes the
data body to be parsed as a NULL-terminated string on all response types
(OKAY, FAIL, INFO, TEXT, DATA).

Note that this matches the default official fastboot tool behavior.

[0] https://android.googlesource.com/platform/system/core/+/master/fastboot/README.md#transport-and-framing

Signed-off-by: Ariel D'Alessandro <ariel.dalessandro@collabora.com>